#193: removed legacy api services
