+++ src/components/Home/FacilityUpdates.jsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { getActiveFacilityUpdates } from "../../services/announcementService";
 import FacilityUpdateCard from "./FacilityUpdateCard";
-import SpinnerWithText from "../common/SpinnerWithText";
 import { toast } from "react-toastify";
 
 class FacilityUpdates extends React.Component {
@@ -14,14 +13,14 @@ class FacilityUpdates extends React.Component {
     this.setState({ showSpinner: true });
 
     try {
-      const { data } = await getActiveFacilityUpdates();
-      let updates = data.results;
+      const { data: res } = await getActiveFacilityUpdates();
+      let updates = res.results;
       updates = updates.map(update => {
         const long_date = update.display_date
         update.display_date = long_date.substring(0, 10);
         return update;
       })
-      this.setState({ updates, showKeySpinner: false });
+      this.setState({ updates, showSpinner: false });
     } catch (ex) {
       toast.error("Failed to load facility updates. Please reload this page.");
       console.log("Failed to load facility updates: " + ex.response.data);
@@ -32,10 +31,7 @@ class FacilityUpdates extends React.Component {
     const { updates, showSpinner } = this.state;
     return (
       <div>
-        {
-          showSpinner ? <SpinnerWithText text={"Loading updates..."} /> : 
-          <FacilityUpdateCard updates={updates} />
-        }
+        <FacilityUpdateCard showSpinner={showSpinner} updates={updates} />
       </div>
     );
   }
